Inline the Sarcophagus query in the Ramses DLMM adapter
Keep the shared funding query in an existing adapter file and reuse it from the CL and Legacy handlers, leaving the final pull request without a new source file.
